requirements.txt
# Core data science
pandas>=2.0.0
numpy>=1.24.0

# Machine learning
scikit-learn>=1.3.0
scikit-surprise>=1.1.3

# API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# Serialization
joblib>=1.3.0

# EDA & visualization
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=8.0.0
httpx>=0.27.0

# Utilities
pyyaml>=6.0
scipy>=1.11.0
